% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{acf_data}
\alias{acf_data}
\title{Estimate autocorrelation}
\usage{
acf_data(x, lag)
}
\arguments{
\item{x}{Single chain.}

\item{lag}{maximum lag. Must be an integer between 1 and 500.}
}
\description{
Estimate autocorrelation
}
