% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{add_parameter}
\alias{add_parameter}
\title{Title}
\usage{
add_parameter(
  x,
  name,
  min = -Inf,
  max = Inf,
  initial_values = NULL,
  blocks = 1L
)
}
\arguments{
\item{x}{The Parameters data.frame to modify or an empty data.frame()}

\item{name}{Parameter name}

\item{min}{Parameter minimum}

\item{max}{Parameter maximum}

\item{initial_values}{Optional vector of initial values. If NULL random starting values will be drawn. Must be between min and max. Must be the same length as the number of chains.}

\item{blocks}{Optional vector of likelihood blocks}
}
\value{
Modified parameters data.frame
}
\description{
Title
}
