% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append_output}
\alias{append_output}
\title{Append output from the cpp mcmc to R6 object}
\usage{
append_output(current, new, phase, theta_names, chain)
}
\arguments{
\item{current}{Any current output (from previous steps)}

\item{new}{New output from cpp mcmc call}

\item{phase}{Phase: tune, burn or sample}

\item{theta_names}{Vector of parameter names}

\item{chain}{Chain}
}
\description{
Append output from the cpp mcmc to R6 object
}
