% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{chain_element}
\alias{chain_element}
\title{Extract names element from nested list}
\usage{
chain_element(x, name)
}
\arguments{
\item{x}{Nested list}

\item{name}{name}
}
\description{
Extract names element from nested list
}
