% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_drjacoby_loaded}
\alias{check_drjacoby_loaded}
\title{Check that drjacoby package has loaded successfully}
\usage{
check_drjacoby_loaded()
}
\description{
Simple function to check that drjacoby package has loaded 
  successfully. Prints "drjacoby loaded successfully!" if so.
}
