% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_checks.R
\name{check_likelihood_compilation}
\alias{check_likelihood_compilation}
\title{Check Cpp loglikelihood function}
\usage{
check_likelihood_compilation(cpp_loglike)
}
\arguments{
\item{cpp_loglike}{Cpp loglikelihood function}
}
\description{
Peforms compilation, return type and argument type checks
}
