% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_checks.R
\name{check_prior_compilation}
\alias{check_prior_compilation}
\title{Check Cpp logprior function}
\usage{
check_prior_compilation(cpp_logprior)
}
\arguments{
\item{cpp_logprior}{Cpp logprior function}
}
\description{
Peforms compilation, return type and argument type checks
}
