% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpp_funciton.R
\name{cpp_function_get}
\alias{cpp_function_get}
\title{Get .cpp function}
\usage{
cpp_function_get(address)
}
\arguments{
\item{address}{Address of .cpp file}
}
\value{
A drJacoby compatible string function
}
\description{
Scrapes an appropriately tagged .cpp function for use in drJacoby. This is convenient as it
facilitates development and testing of a .cpp likelihood or prior via standard Rcpp::sourceCpp.
Once ready the same .cpp file can be imported for use in drJacoby using \code{cpp_function_get()}.
To allow this the user must add two tags. The first, \code{// drjacoby_function_start}, should
be on the line immediately before the start of the function and the second, \code{// drjacoby_function_end},
on the line immediately after the function. Only one function should be tagged in this manner per file.
}
