% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_template.R
\name{cpp_template}
\alias{cpp_template}
\title{Create template for cpp}
\usage{
cpp_template(save_as)
}
\arguments{
\item{save_as}{of file to create, relative to root of active project.}
}
\description{
Create template for cpp
}
