% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_utils.R
\name{create_chain_phase_list}
\alias{create_chain_phase_list}
\title{Create nested list of phases in chains}
\usage{
create_chain_phase_list(chains, base)
}
\arguments{
\item{chains}{Numebr of chains}

\item{base}{element at base of list}
}
\description{
Create nested list of phases in chains
}
