% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_cor_mat_plot}
\alias{create_cor_mat_plot}
\title{Plot posterior correlation matrix}
\usage{
create_cor_mat_plot(output_df, show, phase, param_names)
}
\arguments{
\item{show}{Vector of parameter names to plot.}

\item{param_names}{Optional vector of names to replace the default parameter names.}
}
\description{
Produces a matrix showing the correlation between all parameters
  from posterior draws.
}
