% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_cor_plot}
\alias{create_cor_plot}
\title{Plot parameter correlation}
\usage{
create_cor_plot(parx, pary, output_df, downsample, phase, chain)
}
\arguments{
\item{downsample}{whether to downsample output to speed up plotting.}

\item{parameter1}{name of parameter first parameter.}

\item{parameter2}{name of parameter second parameter.}
}
\description{
Plots correlation between two parameters
}
