% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_credible_plot}
\alias{create_credible_plot}
\title{Plot 95\% credible intervals}
\usage{
create_credible_plot(output_df, show, phase, param_names)
}
\arguments{
\item{show}{vector of parameter names to plot.}

\item{param_names}{optional vector of names to replace the default parameter names.}
}
\description{
Plots posterior 95\% credible intervals over specified set of
  parameters (defauls to all parameters).
}
