% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_par_plot}
\alias{create_par_plot}
\title{Plot parameter estimates}
\usage{
create_par_plot(par, output_df, lag, downsample, phase, chain, return_elements)
}
\arguments{
\item{lag}{maximum lag. Must be an integer between 1 and 500.}

\item{downsample}{boolean. Whether to downsample chain to make plotting more
efficient.}

\item{show}{optional vector of parameter names to plot. Parameters matching
show will be included.}

\item{hide}{optional vector of parameter names to filter out. Parameters
matching hide will be hidden.}

\item{display}{boolean. Whether to show plots, if \code{FALSE} then plotting
objects are returned without displaying.}
}
\description{
Produce a series of plots corresponding to each parameter,
  including the raw trace, the posterior histogram and an autocorrelation
  plot. Plotting objects can be cycled through interactively, or can be
  returned as an object allowing them to be viewed/edited by the user.
}
