% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_utils.R
\name{create_proposal_sd}
\alias{create_proposal_sd}
\title{Create list containing proposal sd matrix for each chain}
\usage{
create_proposal_sd(chains, rungs, n_par, init_sd = 0.1)
}
\arguments{
\item{chains}{Number of chains}

\item{rungs}{Number of rungs}

\item{n_par}{Number of parameters}

\item{init_sd}{Initial value}
}
\description{
Create list containing proposal sd matrix for each chain
}
