% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_v4.R
\name{define_default}
\alias{define_default}
\title{Replace with default values}
\usage{
define_default(x, default)
}
\arguments{
\item{x}{object to define default.}

\item{default}{default value.}
}
\description{
If NULL then replace with chosen value, otherwise keep original
  value.
}
