% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_params}
\alias{define_params}
\title{Define parameters dataframe}
\usage{
define_params(...)
}
\arguments{
\item{...}{define parameters one at a time, for each one defining four
elements: name, min, max, init.}
}
\description{
Provides a convenient way of defining parameters in the format
  required by \code{run_mcmc()}.
}
