% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\name{dj}
\alias{dj}
\title{R6 class representing a mcmc object}
\description{
Create a new dj object to run an mcmc
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{error_debug}}{Diagnostic output in the case of mcmc failure}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dj-new}{\code{dj$new()}}
\item \href{#method-dj-print}{\code{dj$print()}}
\item \href{#method-dj-tune}{\code{dj$tune()}}
\item \href{#method-dj-burn}{\code{dj$burn()}}
\item \href{#method-dj-sample}{\code{dj$sample()}}
\item \href{#method-dj-output}{\code{dj$output()}}
\item \href{#method-dj-acceptance_rate}{\code{dj$acceptance_rate()}}
\item \href{#method-dj-dic}{\code{dj$dic()}}
\item \href{#method-dj-ess}{\code{dj$ess()}}
\item \href{#method-dj-rhat}{\code{dj$rhat()}}
\item \href{#method-dj-timing}{\code{dj$timing()}}
\item \href{#method-dj-plot_par}{\code{dj$plot_par()}}
\item \href{#method-dj-plot_cor}{\code{dj$plot_cor()}}
\item \href{#method-dj-plot_credible}{\code{dj$plot_credible()}}
\item \href{#method-dj-plot_cor_mat}{\code{dj$plot_cor_mat()}}
\item \href{#method-dj-clone}{\code{dj$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-new"></a>}}
\if{latex}{\out{\hypertarget{method-dj-new}{}}}
\subsection{Method \code{new()}}{
Initialise the dj object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$new(
  data,
  df_params,
  loglikelihood,
  logprior,
  chains = 1L,
  misc = list(),
  seed = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a named list or data frame or data values.}

\item{\code{df_params}}{a data.frame of parameters (see \code{?define_params}).}

\item{\code{chains}}{the number of independent replicates of the MCMC to run.}

\item{\code{misc}}{optional list object passed to likelihood and prior. This can be
useful for passing values that are not strictly data, for example passing a
lookup table to the prior function.}

\item{\code{seed}}{Seed for reproducible random number generation. Must be an integer}

\item{\code{loglike, logprior}}{the log-likelihood and log-prior functions used in
the MCMC. Can either be passed in as R functions (not in quotes), or as
character strings naming compiled cpp11 functions (in quotes).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `dj` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-print"></a>}}
\if{latex}{\out{\hypertarget{method-dj-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-tune"></a>}}
\if{latex}{\out{\hypertarget{method-dj-tune}{}}}
\subsection{Method \code{tune()}}{
Run tuning. This will dynamically select the optimum rung number and beta
schedule. Proposal standard deviations will also be tuned in this phase but
target acceptance rates and chain convergence are not guaranteed  - run
\code{$burn()} to burn in further.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$tune(
  iterations,
  beta = seq(1, 0, -0.1),
  max_rungs = 100,
  target_acceptance = 0.44,
  silent = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iterations}}{Number of tuning iterations to run}

\item{\code{beta}}{Initial beta schedule}

\item{\code{max_rungs}}{The maximum number of rungs}

\item{\code{target_acceptance}}{Target acceptance rate}

\item{\code{silent}}{print progress (boolean)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-burn"></a>}}
\if{latex}{\out{\hypertarget{method-dj-burn}{}}}
\subsection{Method \code{burn()}}{
Run burn in. Runs the burn in phase of the MCMC to allow where 
proposal standard deviations to be tuned towards a target acceptance rate
and chains to converge.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$burn(iterations, target_acceptance = 0.44, silent = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iterations}}{Number of tuning iterations to run}

\item{\code{target_acceptance}}{Target acceptance rate}

\item{\code{silent}}{print progress (boolean)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-sample"></a>}}
\if{latex}{\out{\hypertarget{method-dj-sample}{}}}
\subsection{Method \code{sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$sample(iterations, silent = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-output"></a>}}
\if{latex}{\out{\hypertarget{method-dj-output}{}}}
\subsection{Method \code{output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$output(phase = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-acceptance_rate"></a>}}
\if{latex}{\out{\hypertarget{method-dj-acceptance_rate}{}}}
\subsection{Method \code{acceptance_rate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$acceptance_rate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-dic"></a>}}
\if{latex}{\out{\hypertarget{method-dj-dic}{}}}
\subsection{Method \code{dic()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$dic()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-ess"></a>}}
\if{latex}{\out{\hypertarget{method-dj-ess}{}}}
\subsection{Method \code{ess()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$ess()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-rhat"></a>}}
\if{latex}{\out{\hypertarget{method-dj-rhat}{}}}
\subsection{Method \code{rhat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$rhat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-timing"></a>}}
\if{latex}{\out{\hypertarget{method-dj-timing}{}}}
\subsection{Method \code{timing()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$timing()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-plot_par"></a>}}
\if{latex}{\out{\hypertarget{method-dj-plot_par}{}}}
\subsection{Method \code{plot_par()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$plot_par(
  par,
  lag = 20,
  downsample = TRUE,
  phase = "sample",
  chain = NULL,
  return_elements = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-plot_cor"></a>}}
\if{latex}{\out{\hypertarget{method-dj-plot_cor}{}}}
\subsection{Method \code{plot_cor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$plot_cor(parx, pary, downsample = TRUE, phase = "sample", chain = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-plot_credible"></a>}}
\if{latex}{\out{\hypertarget{method-dj-plot_credible}{}}}
\subsection{Method \code{plot_credible()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$plot_credible(
  show = NULL,
  phase = "sample",
  chain = NULL,
  param_names = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-plot_cor_mat"></a>}}
\if{latex}{\out{\hypertarget{method-dj-plot_cor_mat}{}}}
\subsection{Method \code{plot_cor_mat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$plot_cor_mat(show, phase = "sample", chain = NULL, param_names = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dj-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
