% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drjacoby.R
\docType{package}
\name{drjacoby}
\alias{drjacoby}
\alias{drjacoby-package}
\title{Flexible Markov Chain Monte Carlo via Reparameterization}
\description{
Flexible Markov chain monte carlo via reparameterization using
  the Jacobean matrix.
}
