% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{drjacoby_cols}
\alias{drjacoby_cols}
\title{Series of standard colours used within the drjacoby package}
\usage{
drjacoby_cols()
}
\description{
Returns a series of standard colours used within the drjacoby
  package.
}
