% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{drjacoby_file}
\alias{drjacoby_file}
\title{Load system file}
\usage{
drjacoby_file(name)
}
\arguments{
\item{name}{the name of a file within the inst/extdata folder.}
}
\description{
Load a file from within the inst/extdata folder of the drjacoby
  package. File extension must be one of .csv or .rds.
}
