% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dummy1}
\alias{dummy1}
\title{Dummy function}
\usage{
dummy1(x = 1:5)
}
\arguments{
\item{x}{vector of values}
}
\description{
Simple test function that demonstrates some of the features of
  this package.
}
\details{
Takes a vector of values, returns the square.
}
\examples{
# Find square of first 100 values
dummy1(1:100)
}
