% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{ess}
\alias{ess}
\title{Effective sample size}
\usage{
ess(x)
}
\arguments{
\item{x}{chain}
}
\value{
The effective sample size of a chain
}
\description{
Estimates effective sample size. 
Note 1: Lag is trunctaed at when autocorrelation < 0.05, 
if -ve autocorrelation occurs it is possible to get ess > N.
Note 2: ESS estimates will be too optimistic for chains that haven’t mixed.
}
