% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{gelman_rubin}
\alias{gelman_rubin}
\title{Gelman-Rubin statistic}
\usage{
gelman_rubin(par_matrix, chains, samples)
}
\arguments{
\item{par_matrix}{Matrix (interations x chains)}

\item{chains}{number of chains}

\item{samples}{number of samples}
}
\value{
Gelman-Rubin statistic
}
\description{
Estimate sthe Gelman-Rubin (rhat) convergence statistic for a single parameter
across multiple chains. Basic method, assuming all chains are of equal length
}
\references{
Gelman, A., and D. B. Rubin. 1992. 
Inference from Iterative Simulation Using Multiple Sequences. 
Statistical Science 7: 457–511.

\url{https://astrostatistics.psu.edu/RLectures/diagnosticsMCMC.pdf}
}
