% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_theta_rungs}
\alias{get_theta_rungs}
\title{Extract theta into list of matrices over rungs}
\usage{
get_theta_rungs(theta_list, param_names, rung_names)
}
\arguments{
\item{theta_list}{List of thetas}

\item{param_names}{Vector of parameter names}

\item{rung_names}{Vector of rung names}
}
\value{
List of matrices
}
\description{
Extract theta into list of matrices over rungs
}
