% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_and_prior_functions.R
\name{loglike_block}
\alias{loglike_block}
\title{Blocked log likelihood function}
\usage{
loglike_block(params, data, misc)
}
\arguments{
\item{params}{parmeter vector}

\item{data}{data list}

\item{misc}{misc list}
}
\description{
Blocked log likelihood function
}
