% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_and_prior_functions.R
\name{loglike_double_well}
\alias{loglike_double_well}
\title{Double well log likelihood function}
\usage{
loglike_double_well(params, data, misc)
}
\arguments{
\item{params}{parmeter vector}

\item{data}{data list}

\item{misc}{misc list}
}
\description{
Double well log likelihood function
}
