% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_and_prior_functions.R
\name{loglike_normal}
\alias{loglike_normal}
\title{Simple normal log likelihood function}
\usage{
loglike_normal(params, data, misc)
}
\arguments{
\item{params}{parmeter vector}

\item{data}{data list}

\item{misc}{misc list}
}
\description{
Simple normal log likelihood function
}
