% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_and_prior_functions.R
\name{loglike_null}
\alias{loglike_null}
\title{Null log likelihood function}
\usage{
loglike_null(params, data, misc)
}
\arguments{
\item{params}{parmeter vector}

\item{data}{data list}

\item{misc}{misc list}
}
\description{
Null log likelihood function
}
