% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_and_prior_functions.R
\name{logprior_null}
\alias{logprior_null}
\title{Null log prior function}
\usage{
logprior_null(params, misc)
}
\arguments{
\item{params}{parmeter vector}

\item{misc}{misc list}
}
\description{
Null log prior function
}
