% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_autocorrelation}
\alias{plot_autocorrelation}
\title{Plot autocorrelation}
\usage{
plot_autocorrelation(x, lag = 20, par = NULL, chain = 1, phase = "sampling")
}
\arguments{
\item{lag}{maximum lag. Must be an integer between 1 and 500.}

\item{par}{vector of parameter names. If \code{NULL} all parameters are
plotted.}
}
\description{
Plot autocorrelation for specified parameters
}
