% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_contour}
\alias{plot_contour}
\title{Contour plot of two parameters}
\usage{
plot_contour(x, parameter1, parameter2, n_levels = 10)
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}}

\item{parameter1, parameter2}{Name of parameters to plot.}

\item{n_levels}{Number of contour levels.}
}
\description{
Produces contour plot from the posterior draws of two
  parameters.
}
