% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cor}
\alias{plot_cor}
\title{Plot parameter correlation}
\usage{
plot_cor(
  x,
  parameter1,
  parameter2,
  downsample = TRUE,
  phase = "sampling",
  chain = NULL
)
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}}

\item{parameter1}{name of parameter first parameter.}

\item{parameter2}{name of parameter second parameter.}

\item{downsample}{whether to downsample output to speed up plotting.}

\item{phase}{which phase to plot. Must be either "burnin" or "sampling".}

\item{chain}{which chain to plot.}
}
\description{
Plots correlation between two parameters
}
