% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_density}
\alias{plot_density}
\title{Produce density plots}
\usage{
plot_density(x, show = NULL, hide = NULL)
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}}

\item{show}{optional vector of parameter names to plot. Parameters matching
show will be included.}

\item{hide}{optional vector of parameter names to filter out. Parameters
matching hide will be hidden.}
}
\description{
Density plots of all parameters. Use \code{show} and \code{hide}
  to be more specific about which parameters to plot.
}
