% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mc_acceptance}
\alias{plot_mc_acceptance}
\title{Plot Metropolis coupling acceptance rates}
\usage{
plot_mc_acceptance(x, chain = NULL, phase = "sampling", x_axis_type = 1)
}
\arguments{
\item{chain}{which chain to plot. If \code{NULL} then plot all chains.}
}
\description{
Plot Metropolis coupling acceptance rates between all rungs.
}
