% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pairs}
\alias{plot_pairs}
\title{Produce scatterplots between multiple parameters}
\usage{
plot_pairs(x, show = NULL, hide = NULL)
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}}

\item{show}{optional vector of parameter names to plot. Parameters matching
show will be included.}

\item{hide}{optional vector of parameter names to filter out. Parameters
matching hide will be hidden.}
}
\description{
Uses \code{ggpairs} function from the \code{GGally} package to
  produce scatterplots between all named parameters.
}
