% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_par}
\alias{plot_par}
\title{Plot parameter estimates}
\usage{
plot_par(
  x,
  show = NULL,
  hide = NULL,
  lag = 20,
  downsample = TRUE,
  phase = "sampling",
  rung = 1,
  chain = "all",
  display = TRUE
)
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}}

\item{show}{Optional character (or vector of characters) to filter parameters by.
Parameters matching show will be included.}

\item{hide}{Optional character (or vector of characters) to filter parameters by.
Parameters matching show will be hidden.}

\item{lag}{Maximum lag. Must be an integer between 1 and 500.}

\item{downsample}{Boolean. Whether to downsample chain to make plotting more
efficient.}

\item{phase}{which phase to plot. Must be either "burnin" or "sampling".}

\item{rung}{Which temperature rung to plot. Defaults to the cold chain.}

\item{chain}{Which chain to plot, e.g. \code{"chain1"}. The default
\code{"all"} plots all chains.}

\item{display}{Boolean. Whether to show plots, if \code{FALSE} then plotting
objects are returned without displaying.}
}
\description{
Produce a series of plots corresponding to each parameter,
  including the raw trace, the posterior histogram and an autocorrelation
  plot. Plotting objects can be cycled through interactively, or can be
  returned as an object allowing them to be viewed/edited by the user.
}
