% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_rung_loglike}
\alias{plot_rung_loglike}
\title{Plot loglikelihood 95\% credible intervals}
\usage{
plot_rung_loglike(x, chain = 1, phase = "sampling", x_axis_type = 1,
  y_axis_type = 1)
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}}

\item{chain}{which chain to plot.}

\item{phase}{which phase to plot. Must be either "burnin" or "sampling".}

\item{x_axis_type}{how to format the x-axis. 1 = integer rungs, 2 = values of
the thermodynamic power.}

\item{y_axis_type}{how to format the y-axis. 1 = raw values, 2 = truncated at
auto-chosen lower limit. 3 = double-log scale.}
}
\description{
Plot loglikelihood 95\% credible intervals.
}
