% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Produce bivariate scatterplot}
\usage{
plot_scatter(
  x,
  parameter1,
  parameter2,
  downsample = TRUE,
  phase = "sampling",
  chain = NULL
)
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}}

\item{parameter1}{name of parameter first parameter.}

\item{parameter2}{name of parameter second parameter.}

\item{downsample}{whether to downsample output to 200 values max to speed up
plotting.}

\item{phase}{which phase to plot. Must be either "burnin" or "sampling".}

\item{chain}{which chain to plot.}
}
\description{
Produces scatterplot between two named parameters.
}
