% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot parameter trace}
\usage{
plot_trace(
  x,
  show = NULL,
  hide = NULL,
  lag = 20,
  downsample = TRUE,
  phase = "sampling",
  chain = NULL,
  display = TRUE
)
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}}

\item{show}{optional vector of parameter names to plot. Parameters matching
show will be included.}

\item{hide}{optional vector of parameter names to filter out. Parameters
matching hide will be hidden.}

\item{lag}{maximum lag. Must be an integer between 1 and 500.}

\item{downsample}{boolean. Whether to downsample chain to make plotting more
efficient.}

\item{phase}{which phase to plot. Must be either "burnin", "sampling" or "both".}

\item{chain}{which chain to plot.}

\item{display}{boolean. Whether to show plots, if \code{FALSE} then plotting
objects are returned without displaying.}
}
\description{
Produce a series of plots corresponding to each parameter,
  including the raw trace, the posterior histogram and an autocorrelation
  plot. Plotting objects can be cycled through interactively, or can be
  returned as an object allowing them to be viewed/edited by the user.
}
