% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_data}
\alias{population_data}
\title{Example population data}
\format{
An object of class \code{data.frame} with 20 rows and 2 columns.
}
\usage{
population_data
}
\description{
Example population data
}
\keyword{datasets}
