% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{set_rhat}
\alias{set_rhat}
\title{Gelman-Rubin statistic: all parameters}
\usage{
set_rhat(output, chains)
}
\arguments{
\item{output}{MCMC output}

\item{chains}{the number of independent replicates of the MCMC chain to run.
If a \code{cluster} object is defined then these chains are run in
parallel, otherwise they are run in serial.}
}
\value{
Vector of Gelman-Rubin statistics
}
\description{
Estimates the Gelman-Rubin (rhat) convergence statistic for a all parameters
}
