% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tune_rungs}
\alias{tune_rungs}
\title{Tune mcmc rungs}
\usage{
tune_rungs(x, target_acceptance = 0.23)
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}, with the MCMC completed
and run with parallel tempering over multiple rungs.}

\item{target_acceptance}{the target acceptance rate between rungs. Higher
values will lead to more rungs being needed.}
}
\description{
Takes a completed mcmc object that has been run with parallel
  tempering, and uses the acceptance rates to recalculate the optimal number
  and distribution of rungs.
}
