% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\name{dj}
\alias{dj}
\title{R6 class representing a mcmc object}
\description{
Create a new dj object to run an mcmc
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{error_debug}}{Diagnostic output in the case of mcmc failure}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dj-new}{\code{dj$new()}}
\item \href{#method-dj-print}{\code{dj$print()}}
\item \href{#method-dj-tune}{\code{dj$tune()}}
\item \href{#method-dj-burn}{\code{dj$burn()}}
\item \href{#method-dj-sample}{\code{dj$sample()}}
\item \href{#method-dj-output}{\code{dj$output()}}
\item \href{#method-dj-acceptance_rate}{\code{dj$acceptance_rate()}}
\item \href{#method-dj-mc_acceptance_rate}{\code{dj$mc_acceptance_rate()}}
\item \href{#method-dj-dic}{\code{dj$dic()}}
\item \href{#method-dj-ess}{\code{dj$ess()}}
\item \href{#method-dj-rhat}{\code{dj$rhat()}}
\item \href{#method-dj-timing}{\code{dj$timing()}}
\item \href{#method-dj-plot_par}{\code{dj$plot_par()}}
\item \href{#method-dj-plot_cor}{\code{dj$plot_cor()}}
\item \href{#method-dj-plot_credible}{\code{dj$plot_credible()}}
\item \href{#method-dj-plot_cor_mat}{\code{dj$plot_cor_mat()}}
\item \href{#method-dj-plot_mc_acceptance_rate}{\code{dj$plot_mc_acceptance_rate()}}
\item \href{#method-dj-plot_tuning_rejection_rate}{\code{dj$plot_tuning_rejection_rate()}}
\item \href{#method-dj-plot_local_communication_barrier}{\code{dj$plot_local_communication_barrier()}}
\item \href{#method-dj-clone}{\code{dj$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-new"></a>}}
\if{latex}{\out{\hypertarget{method-dj-new}{}}}
\subsection{Method \code{new()}}{
Initialise the dj object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$new(data, df_params, loglikelihood, logprior, chains = 1L, misc = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a named list or data frame or data values.}

\item{\code{df_params}}{a data.frame of parameters (see \code{?define_params}).}

\item{\code{loglikelihood}}{the log-likelihood function used in
the MCMC. Can either be passed in as R functions (not in quotes), or as
character strings naming compiled cpp11 functions (in quotes).}

\item{\code{logprior}}{the log-prior function used in
the MCMC. Can either be passed in as R functions (not in quotes), or as
character strings naming compiled cpp11 functions (in quotes).}

\item{\code{chains}}{the number of independent replicates of the MCMC to run.}

\item{\code{misc}}{optional list object passed to likelihood and prior. This can be
useful for passing values that are not strictly data, for example passing a
lookup table to the prior function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `dj` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-print"></a>}}
\if{latex}{\out{\hypertarget{method-dj-print}{}}}
\subsection{Method \code{print()}}{
Print mcmc object summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-tune"></a>}}
\if{latex}{\out{\hypertarget{method-dj-tune}{}}}
\subsection{Method \code{tune()}}{
Run tuning. This will dynamically select the optimum rung number and beta
schedule. Proposal standard deviations will also be tuned in this phase but
target acceptance rates and chain convergence are not guaranteed  - run
\code{$burn()} to burn in further.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$tune(
  target_rung_acceptance = 0.5,
  n_rungs = 50,
  iterations = 1000,
  initial_beta = NULL,
  swap = 2L,
  target_acceptance = 0.44,
  silent = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_rung_acceptance}}{The target (average) acceptance rate between rungs}

\item{\code{n_rungs}}{Number of rungs to run for tuning. This should be high enough such that no
links between neighbouring chains have 0% acceptance of swaps. One tuned, the optimised number
of rungs will likely be lower, leading to faster run times for \code{$burn()} and \code{$sample()}}

\item{\code{iterations}}{Number of tuning iterations to run}

\item{\code{initial_beta}}{If prior knowledge of beta exists, the user can override the default tuning beta
defined by \code{n_rungs}. This should be a vector decreaseing from 1 to 0.}

\item{\code{swap}}{integer 0 = no swapping, 1 = standard swapping, 2 = super cool new swapping}

\item{\code{target_acceptance}}{Target acceptance rate}

\item{\code{silent}}{print progress (boolean)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-burn"></a>}}
\if{latex}{\out{\hypertarget{method-dj-burn}{}}}
\subsection{Method \code{burn()}}{
Run burn in. Runs the burn in phase of the MCMC to allow where 
proposal standard deviations to be tuned towards a target acceptance rate
and chains to converge.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$burn(iterations, target_acceptance = 0.44, silent = FALSE, cluster = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iterations}}{Number of burn-in iterations to run}

\item{\code{target_acceptance}}{Target acceptance rate}

\item{\code{silent}}{print progress (boolean)}

\item{\code{cluster}}{parallel cluster object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-sample"></a>}}
\if{latex}{\out{\hypertarget{method-dj-sample}{}}}
\subsection{Method \code{sample()}}{
Run sampling in. Runs the sampling phase of the MCMC.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$sample(iterations, silent = FALSE, cluster = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iterations}}{Number of sampling iterations to run}

\item{\code{silent}}{print progress (boolean)}

\item{\code{cluster}}{parallel cluster object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-output"></a>}}
\if{latex}{\out{\hypertarget{method-dj-output}{}}}
\subsection{Method \code{output()}}{
Get mcmc output data.frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$output(phase = NULL, chain = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phase}}{optional phase selection, can be a vector chosen from "tune", "burn", "sample"}

\item{\code{chain}}{option chain(s) selection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-acceptance_rate"></a>}}
\if{latex}{\out{\hypertarget{method-dj-acceptance_rate}{}}}
\subsection{Method \code{acceptance_rate()}}{
Get acceptance rates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$acceptance_rate(phase = "sample", chain = NULL, rung = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phase}}{optional phase selection, can be a vector chosen from "tune", "burn", "sample"}

\item{\code{chain}}{option chain(s) selection}

\item{\code{rung}}{optional rung selection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-mc_acceptance_rate"></a>}}
\if{latex}{\out{\hypertarget{method-dj-mc_acceptance_rate}{}}}
\subsection{Method \code{mc_acceptance_rate()}}{
Get acceptance rates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$mc_acceptance_rate(phase = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phase}}{optional phase selection, can be a vector chosen from "tune", "burn", "sample"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-dic"></a>}}
\if{latex}{\out{\hypertarget{method-dj-dic}{}}}
\subsection{Method \code{dic()}}{
Get DIC estimate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$dic()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-ess"></a>}}
\if{latex}{\out{\hypertarget{method-dj-ess}{}}}
\subsection{Method \code{ess()}}{
Get effective sample size estimates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$ess()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-rhat"></a>}}
\if{latex}{\out{\hypertarget{method-dj-rhat}{}}}
\subsection{Method \code{rhat()}}{
Get rhat estimates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$rhat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-timing"></a>}}
\if{latex}{\out{\hypertarget{method-dj-timing}{}}}
\subsection{Method \code{timing()}}{
Get run-time data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$timing()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-plot_par"></a>}}
\if{latex}{\out{\hypertarget{method-dj-plot_par}{}}}
\subsection{Method \code{plot_par()}}{
Produce a parameter plot of the named parameter,
  including the raw trace, the posterior histogram and an autocorrelation
  plot. The combined plot or a list of plot elements can be returned.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$plot_par(
  par,
  lag = 20,
  downsample = TRUE,
  phase = NULL,
  chain = NULL,
  return_elements = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{par}}{The name of the parameter}

\item{\code{lag}}{Maximum lag. Must be an integer between 1 and 500.}

\item{\code{downsample}}{boolean. Whether to downsample chain to make plotting more
efficient.}

\item{\code{phase}}{Optional selection of phases, can be from: tune, burn and sample}

\item{\code{chain}}{Optional selection of chains}

\item{\code{return_elements}}{boolean. If \code{TRUE} a list of plotting objects 
are returned without displaying.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-plot_cor"></a>}}
\if{latex}{\out{\hypertarget{method-dj-plot_cor}{}}}
\subsection{Method \code{plot_cor()}}{
Plots the correlation between two parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$plot_cor(parx, pary, downsample = TRUE, phase = "sample", chain = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parx}}{Name of parameter 1}

\item{\code{pary}}{Name of parameter 2}

\item{\code{downsample}}{boolean. Whether to downsample chain to make plotting more
efficient.}

\item{\code{phase}}{Optional selection of phases, can be from: tune, burn and sample}

\item{\code{chain}}{Optional selection of chains}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-plot_credible"></a>}}
\if{latex}{\out{\hypertarget{method-dj-plot_credible}{}}}
\subsection{Method \code{plot_credible()}}{
Plots posterior 95\% credible intervals over specified set of
  parameters (defauls to all parameters).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$plot_credible(
  pars = NULL,
  phase = "sample",
  chain = NULL,
  param_names = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pars}}{Vector of parameter names}

\item{\code{phase}}{Optional selection of phases, can be from: tune, burn and sample}

\item{\code{chain}}{Optional selection of chains}

\item{\code{param_names}}{Optional vector of prameter names for plotting labels}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-plot_cor_mat"></a>}}
\if{latex}{\out{\hypertarget{method-dj-plot_cor_mat}{}}}
\subsection{Method \code{plot_cor_mat()}}{
Produces a matrix showing the correlation between all parameters
  from posterior draws.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$plot_cor_mat(
  pars = NULL,
  phase = "sample",
  chain = NULL,
  param_names = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pars}}{Vector of parameter names}

\item{\code{phase}}{Optional selection of phases, can be from: tune, burn and sample}

\item{\code{chain}}{Optional selection of chains}

\item{\code{param_names}}{Optional vector of prameter names for plotting labels}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-plot_mc_acceptance_rate"></a>}}
\if{latex}{\out{\hypertarget{method-dj-plot_mc_acceptance_rate}{}}}
\subsection{Method \code{plot_mc_acceptance_rate()}}{
Plots the acceptance rate between rungs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$plot_mc_acceptance_rate(phase = "sample", x_axis_type = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phase}}{Optional selection of phases, can be from: tune, burn or sample}

\item{\code{x_axis_type}}{how to format the x-axis. 1 = integer rungs, 2 = values of
the thermodynamic power.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-plot_tuning_rejection_rate"></a>}}
\if{latex}{\out{\hypertarget{method-dj-plot_tuning_rejection_rate}{}}}
\subsection{Method \code{plot_tuning_rejection_rate()}}{
Plots the cumulative rejection rate between rungs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$plot_tuning_rejection_rate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-plot_local_communication_barrier"></a>}}
\if{latex}{\out{\hypertarget{method-dj-plot_local_communication_barrier}{}}}
\subsection{Method \code{plot_local_communication_barrier()}}{
Plots the local communication barrier between rungs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$plot_local_communication_barrier()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dj-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dj-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
