% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included_likelihood_prior.R
\name{loglike_double_well}
\alias{loglike_double_well}
\title{Double well log likelihood function}
\usage{
loglike_double_well(params, data, misc)
}
\description{
Double well log likelihood function
}
