% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_and_prior_functions.R
\name{logprior_return}
\alias{logprior_return}
\title{Log prior function for return prior check}
\usage{
logprior_return(params, misc)
}
\description{
Log prior function for return prior check
}
