% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_autocorrelation}
\alias{plot_autocorrelation}
\title{Plot autocorrelation}
\usage{
plot_autocorrelation(x, lag = 20, par = NULL, chain = 1,
  phase = "sampling")
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}}

\item{lag}{Maximum lag. Must be an integer between 20 and 500.}

\item{par}{Vector of parameter names. If NULL all parameters are plotted}

\item{chain}{which chain to plot.}

\item{phase}{which phase to plot. Must be either "burnin" or "sampling".}
}
\description{
Plot autocorrelation for specified parameters
}
