% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cor}
\alias{plot_cor}
\title{Plot parameter correlation}
\usage{
plot_cor(x, parameter1, parameter2, downsample = TRUE,
  phase = "sampling", rung = 1)
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}}

\item{parameter1}{Name of parameter first parameter.}

\item{parameter2}{Name of parameter second parameter.}

\item{downsample}{Downsample chain for efficiency}

\item{phase}{which phase to plot. Must be either "burnin" or "sampling".}
}
\description{
Plots correlation between two parameters
}
