% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cor_mat}
\alias{plot_cor_mat}
\title{Plot posterior correlation matrix}
\usage{
plot_cor_mat(x, show = NULL, phase = "sampling", param_names = NULL)
}
\arguments{
\item{show}{Vector of parameter names to plot.}

\item{param_names}{Optional vector of names to replace the default parameter names.}
}
\description{
Produces a matrix showing the correlation between all parameters
  from posterior draws.
}
