% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mc_acceptance}
\alias{plot_mc_acceptance}
\title{Plot Metropolis Coupling Acceptance}
\usage{
plot_mc_acceptance(x, chain = 1, phase = "sampling")
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}}

\item{chain}{which chain to plot.}

\item{phase}{which phase to plot. Must be either "burnin" or "sampling".}
}
\description{
Plot Metropolis Coupling Acceptance.
}
