% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mc_acceptance}
\alias{plot_mc_acceptance}
\title{Plot Metropolis coupling acceptance rates}
\usage{
plot_mc_acceptance(x, chain = "all", phase = "sampling", x_axis_type = 1)
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}}

\item{chain}{which chain to plot.}

\item{phase}{which phase to plot. Must be either "burnin" or "sampling".}

\item{x_axis_type}{how to format the x-axis. 1 = integer rungs, 2 = values of
the thermodynamic power.}
}
\description{
Plot Metropolis coupling acceptance rates between all rungs.
}
