% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_par}
\alias{plot_par}
\title{Plot parameter estimates}
\usage{
plot_par(x, phase = "sampling", show = NULL, hide = NULL, lag = 20,
  downsample = TRUE)
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}}

\item{phase}{which phase to plot. Must be either "burnin" or "sampling".}

\item{show}{Optional character (or vector of characters) to filter parameters by.
Parameters matching show will be included.}

\item{hide}{Optional character (or vector of characters) to filter parameters by.
Parameters matching show will be hidden.}

\item{lag}{Maximum lag. Must be an integer between 20 and 500}

\item{downsample}{Downsample chain for efficiency}
}
\description{
Plot parameter estimates
}
