% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_par}
\alias{plot_par}
\title{Plot parameter estimates}
\usage{
plot_par(x, parameter)
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}}

\item{parameter}{Name of parameter}
}
\description{
Plot parameter estimates
}
