% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_data}
\alias{population_data}
\title{Population data.}
\format{
A data frame with 20 rows and 2 variables:
\describe{
  \item{pop}{population size}
  \item{time}{time}
  ...
}
}
\usage{
population_data
}
\description{
Example population growth data.
}
\keyword{datasets}
