% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dj11_run.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run drjacoby MCM}
\usage{
run_mcmc(
  data,
  df_params,
  loglike,
  logprior,
  burnin,
  samples,
  misc = list(),
  rungs = 1L,
  chains = 1,
  beta_manual = NULL,
  alpha = 1,
  target_acceptance = 0.44,
  cluster = NULL,
  coupling_on = TRUE,
  progress = TRUE
)
}
\description{
Run drjacoby MCM
}
