% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run MCMC}
\usage{
run_mcmc(data, df_params, loglike, logprior, df_burnin = NULL,
  samples = 10000, rungs = 1, chains = 1, coupling_on = TRUE,
  GTI_pow = 3, cluster = NULL, pb_markdown = FALSE, silent = FALSE)
}
\arguments{
\item{data}{TODO.}

\item{df_params}{a dataframe of parameters. Must contain the following
elements:
\itemize{
  \item \code{name} - the parameter name.
  \item \code{min} - the minimum value of the parameter. \code{-Inf} is
  allowed.
  \item \code{max} - the maximum value of the parameter. \code{Inf} is
  allowed.
  \item \code{init} - the initial value of the parameter.
}}

\item{loglike}{TODO.}

\item{logprior}{TODO.}

\item{df_burnin}{a dataframe specifying the parameters of each phase of
burn-in. Must contain the following elements:
\itemize{
  \item \code{iterations} - the number of MCMC iterations this phase.
  \item \code{prop_method} - integer flag specifying which method to use
  when proposing new values this phase. 1 = independent proposal per
  parameter, 2 = multivariate proposal over all parameters.
  \item \code{bw_update} - boolean, whether to update the proposal
  bandwidth dynamically via Robbins-Monro this phase. If \code{TRUE} then
  for the independent proposal method this means updating a different
  bandwidth per parameter, for the multivariate proposal method this means
  updating a single bandwidth for all parameters, scaled by a covariance
  matrix.
  \item \code{bw_reset} - boolean, whether the Robbins-Monro index should
  be reset at the beginning of this phase. Resetting the index makes the
  proposal bandwidths more free to move to new values.
  \item \code{cov_recalc} boolean, whether the multivariate proposal
  covariance matrix should be re-calculated from the posterior draws at the
  end of this burn-in phase.
}
If \code{NULL} then a default dataframe is used (see details).}

\item{samples}{the number of sampling iterations.}

\item{rungs}{the number of temperature rungs used in Metropolis coupling (see
\code{coupling_on}).}

\item{chains}{the number of independent replicates of the MCMC chain to run.
If a \code{cluster} object is defined then these chains are run in
parallel, otherwise they are run in serial.}

\item{coupling_on}{whether to implement Metropolis-coupling over temperature 
rungs.}

\item{GTI_pow}{the power used in the generalised thermodynamic integration 
method. Must be greater than 1.0.}

\item{cluster}{option to pass in a cluster environment, allowing chains to be
run in parallel (see package "parallel").}

\item{pb_markdown}{whether to run progress bars in markdown mode, meaning
they are only updated when they reach 100% to avoid large amounts of output
being printed to markdown files.}

\item{silent}{whether to suppress all console output.}
}
\description{
Run MCMC.
}
\details{
TODO - explain df_burnin defaults.
}
