% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run drjacoby MCMC}
\usage{
run_mcmc(
  data,
  df_params,
  loglike,
  logprior,
  burnin = 1000,
  samples = 10000,
  rungs = 1,
  chains = 5,
  coupling_on = TRUE,
  GTI_pow = 3,
  cluster = NULL,
  pb_markdown = FALSE,
  silent = FALSE
)
}
\arguments{
\item{data}{a vector of data values. When using C++ versions of the
likelihood and/or prior these values are treated internally as doubles, so
while integer and boolean values can be used, keep in mind that these will
be recast as doubles in the likelihood (i.e. \code{TRUE = 1.0}).}

\item{df_params}{a dataframe of parameters. Must contain the following
elements:
\itemize{
  \item \code{name} - the parameter name.
  \item \code{min} - the minimum value of the parameter. \code{-Inf} is
  allowed.
  \item \code{max} - the maximum value of the parameter. \code{Inf} is
  allowed.
  \item \code{init} - the initial value of the parameter.
}}

\item{loglike}{TODO.}

\item{logprior}{TODO.}

\item{burnin}{the number of burn-in iterations.}

\item{samples}{the number of sampling iterations.}

\item{rungs}{the number of temperature rungs used in Metropolis coupling (see
\code{coupling_on}).}

\item{chains}{the number of independent replicates of the MCMC chain to run.
If a \code{cluster} object is defined then these chains are run in
parallel, otherwise they are run in serial.}

\item{coupling_on}{whether to implement Metropolis-coupling over temperature 
rungs.}

\item{GTI_pow}{the power used in the generalised thermodynamic integration 
method.}

\item{cluster}{option to pass in a cluster environment, allowing chains to be
run in parallel (see package "parallel").}

\item{pb_markdown}{whether to run progress bars in markdown mode, meaning
they are only updated when they reach 100% to avoid large amounts of output
being printed to markdown files.}

\item{silent}{whether to suppress all console output.}
}
\description{
Run flexible MCMC through drjacoby.
}
