% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{sample_chains}
\alias{sample_chains}
\title{Sample N draws from all available chains}
\usage{
sample_chains(x, sample_n)
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}}

\item{sample_n}{An integer number of samples}
}
\value{
A dataframe of posterior samples
}
\description{
Sample N draws from all available chains
}
